requirements.txt
cachetools==5.3.2
et-xmlfile==1.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.1
oauthlib==3.2.2
openpyxl==3.1.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
StrEnum==0.4.15
